Reorganize Bazel functionality into separate InjectionBazel module
- Create new InjectionBazel module with dedicated directory structure
- Move all Bazel-related files: BazelAQueryParser, BazelActionQueryHandler, BazelInterface
- Add platform constraints and proper dependency management in Package.swift
- Wrap all Bazel code in #if DEBUG || \!SWIFT_PACKAGE conditional compilation
- Move LiteParser protocol to InjectionBazel module for proper abstraction
- Create separate InjectionBazelTests module with cleaned test suite
- Update InjectionLite imports to reference InjectionBazel module
- Add DLKitD dependency for InjectionBazel module

This modular structure keeps Bazel functionality optional and well-organized
while maintaining clean separation of concerns between core InjectionLite
and Bazel-specific implementation.
