--- Tests/InjectionLiteTests/BazelTests.swift
@@ -1,262 +1,0 @@
-//
-//  BazelTests.swift
-//  InjectionLiteTests
-//
-//  Tests for Bazel AQuery integration functionality
-//
-
-import XCTest
-@testable import InjectionLite
-
-final class BazelTests: XCTestCase {
-    
-    private var tempWorkspaceURL: URL!
-    private var workspacePath: String!
-    
-    override func setUp() {
-        super.setUp()
-        
-        // Create temporary Bazel workspace for testing
-        tempWorkspaceURL = URL(fileURLWithPath: NSTemporaryDirectory())
-            .appendingPathComponent("BazelTestWorkspace_\(UUID().uuidString)")
-        
-        try! FileManager.default.createDirectory(at: tempWorkspaceURL, 
-                                               withIntermediateDirectories: true)
-        workspacePath = tempWorkspaceURL.path
-    }
-    
-    override func tearDown() {
-        // Clean up temporary workspace
-        if let tempWorkspaceURL = tempWorkspaceURL {
-            try? FileManager.default.removeItem(at: tempWorkspaceURL)
-        }
-        super.tearDown()
-    }
-    
-    // MARK: - Workspace Detection Tests
-    
-    func testBazelWorkspaceDetectionWithMODULE() {
-        // Create MODULE.bazel file
-        let moduleFile = tempWorkspaceURL.appendingPathComponent("MODULE.bazel")
-        XCTAssertNoThrow(try "module(name = \"test\")".write(to: moduleFile, atomically: true, encoding: .utf8))
-        
-        // Test detection
-        XCTAssertTrue(BazelInterface.isBazelWorkspace(containing: workspacePath))
-        
-        let foundRoot = BazelInterface.findWorkspaceRoot(containing: workspacePath)
-        XCTAssertEqual(foundRoot, workspacePath)
-    }
-    
-    func testBazelWorkspaceDetectionWithWORKSPACE() {
-        // Create WORKSPACE file
-        let workspaceFile = tempWorkspaceURL.appendingPathComponent("WORKSPACE")
-        try! "workspace(name = \"test\")".write(to: workspaceFile, atomically: true, encoding: .utf8)
-        
-        // Test detection
-        XCTAssertTrue(BazelInterface.isBazelWorkspace(containing: workspacePath))
-        
-        let foundRoot = BazelInterface.findWorkspaceRoot(containing: workspacePath)
-        XCTAssertEqual(foundRoot, workspacePath)
-    }
-    
-    func testBazelWorkspaceDetectionWithWORKSPACEBazel() {
-        // Create WORKSPACE.bazel file
-        let workspaceFile = tempWorkspaceURL.appendingPathComponent("WORKSPACE.bazel")
-        try! "workspace(name = \"test\")".write(to: workspaceFile, atomically: true, encoding: .utf8)
-        
-        // Test detection
-        XCTAssertTrue(BazelInterface.isBazelWorkspace(containing: workspacePath))
-        
-        let foundRoot = BazelInterface.findWorkspaceRoot(containing: workspacePath)
-        XCTAssertEqual(foundRoot, workspacePath)
-    }
-    
-    func testBazelWorkspaceDetectionWithPlainMODULE() {
-        // Create plain MODULE file (without .bazel extension)
-        let moduleFile = tempWorkspaceURL.appendingPathComponent("MODULE")
-        try! "module(name = \"test\")".write(to: moduleFile, atomically: true, encoding: .utf8)
-        
-        // Test detection
-        XCTAssertTrue(BazelInterface.isBazelWorkspace(containing: workspacePath))
-        
-        let foundRoot = BazelInterface.findWorkspaceRoot(containing: workspacePath)
-        XCTAssertEqual(foundRoot, workspacePath)
-    }
-    
-    func testNonBazelWorkspaceDetection() {
-        // No Bazel files - should not be detected as Bazel workspace
-        XCTAssertFalse(BazelInterface.isBazelWorkspace(containing: workspacePath))
-        XCTAssertNil(BazelInterface.findWorkspaceRoot(containing: workspacePath))
-    }
-    
-    func testNestedWorkspaceDetection() {
-        // Create nested directory structure
-        let nestedDir = tempWorkspaceURL.appendingPathComponent("src/main/swift")
-        try! FileManager.default.createDirectory(at: nestedDir, withIntermediateDirectories: true)
-        
-        // Create MODULE.bazel at root
-        let moduleFile = tempWorkspaceURL.appendingPathComponent("MODULE.bazel")
-        try! "module(name = \"test\")".write(to: moduleFile, atomically: true, encoding: .utf8)
-        
-        // Test detection from nested directory
-        let nestedPath = nestedDir.path
-        XCTAssertTrue(BazelInterface.isBazelWorkspace(containing: nestedPath))
-        
-        let foundRoot = BazelInterface.findWorkspaceRoot(containing: nestedPath)
-        XCTAssertEqual(foundRoot, workspacePath)
-    }
-    
-    // MARK: - BazelInterface Tests
-    
-    func testBazelInterfaceInitialization() {
-        // Create valid workspace
-        let moduleFile = tempWorkspaceURL.appendingPathComponent("MODULE.bazel")
-        try! "module(name = \"test\")".write(to: moduleFile, atomically: true, encoding: .utf8)
-        
-        // Test initialization
-        XCTAssertNoThrow(try BazelInterface(workspaceRoot: workspacePath, bazelExecutable: "echo"))
-    }
-    
-    func testBazelInterfaceInitializationFailure() {
-        // No Bazel files - should fail initialization
-        XCTAssertThrowsError(try BazelInterface(workspaceRoot: workspacePath)) { error in
-            XCTAssertTrue(error is BazelError)
-            if case BazelError.workspaceNotFound(let path) = error {
-                XCTAssertEqual(path, workspacePath)
-            } else {
-                XCTFail("Expected workspaceNotFound error")
-            }
-        }
-    }
-    
-    
-    // MARK: - Recompiler Integration Tests
-    
-    func testRecompilerBazelDetection() {
-        // Create Bazel workspace
-        let moduleFile = tempWorkspaceURL.appendingPathComponent("MODULE.bazel")
-        try! "module(name = \"test\")".write(to: moduleFile, atomically: true, encoding: .utf8)
-        
-        // Create source file
-        let sourceFile = tempWorkspaceURL.appendingPathComponent("test.swift")
-        let sourcePath = sourceFile.path
-        try! "print(\"test\")".write(to: sourceFile, atomically: true, encoding: .utf8)
-        
-        let recompiler = Recompiler()
-        let parser = recompiler.parser(forProjectContaining: sourcePath)
-        
-        // Should return BazelAQueryParser for Bazel projects
-        XCTAssertTrue(parser is BazelAQueryParser)
-    }
-    
-    func testRecompilerXcodeDetection() {
-        // Create non-Bazel source file
-        let sourceFile = tempWorkspaceURL.appendingPathComponent("test.swift")
-        let sourcePath = sourceFile.path
-        try! "print(\"test\")".write(to: sourceFile, atomically: true, encoding: .utf8)
-        
-        let recompiler = Recompiler()
-        let parser = recompiler.parser(forProjectContaining: sourcePath)
-        
-        // Should return LogParser for non-Bazel projects
-        XCTAssertTrue(parser is LogParser)
-    }
-    
-    // MARK: - Error Handling Tests
-    
-    func testBazelErrorDescriptions() {
-        let errors = [
-            BazelError.workspaceNotFound("/path/to/workspace"),
-            BazelError.bazelNotFound,
-            BazelError.queryFailed("query error"),
-            BazelError.buildFailed("build error"),
-            BazelError.targetNotFound("//target:name"),
-            BazelError.invalidPath("/invalid/path"),
-            BazelError.pathResolutionFailed("/path")
-        ]
-        
-        for error in errors {
-            let description = error.description
-            XCTAssertFalse(description.isEmpty, "Error description should not be empty")
-            XCTAssertTrue(description.count > 10, "Error description should be meaningful")
-        }
-    }
-    
-    func testBazelPathErrorDescriptions() {
-        let errors = [
-            BazelPathError.invalidPath("/path"),
-            BazelPathError.packageNotFound("/path"),
-            BazelPathError.workspaceNotFound,
-            BazelPathError.buildFileNotFound("/path"),
-            BazelPathError.pathOutsideWorkspace("/path"),
-            BazelPathError.labelGenerationFailed("/path")
-        ]
-        
-        for error in errors {
-            let description = error.description
-            XCTAssertFalse(description.isEmpty, "Error description should not be empty")
-            XCTAssertTrue(description.count > 10, "Error description should be meaningful")
-        }
-    }
-    
-    // MARK: - Command Cleaning Tests
-    
-    func testBazelCommandCleaning() {
-        // Create valid workspace
-        let moduleFile = tempWorkspaceURL.appendingPathComponent("MODULE.bazel")
-        try! "module(name = \"test\")".write(to: moduleFile, atomically: true, encoding: .utf8)
-        
-        XCTAssertNoThrow(try {
-            let parser = try BazelAQueryParser(workspaceRoot: workspacePath)
-            
-            // Mock command with flags that should be removed (matching real Bazel output)
-            let originalCommand = """
-                /usr/bin/swiftc -module-name TestModule \
-                -Xfrontend \
-                -const-gather-protocols-file \
-                -Xfrontend \
-                /tmp/const_protocols_to_gather.json \
-                -emit-const-values-path \
-                bazel-out/ios_sim_arm64-dbg/bin/TestModule.swiftconstvalues \
-                -emit-module-path \
-                bazel-out/ios_sim_arm64-dbg/bin/TestModule.swiftmodule \
-                -index-store-path \
-                bazel-out/ios_sim_arm64-dbg/bin/TestModule.indexstore \
-                -index-ignore-system-modules \
-                '-Xwrapped-swift=-global-index-store-import-path=bazel-out/_global_index_store' \
-                -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator18.1.sdk \
-                -o output.o input.swift
-                """
-            
-            // Test the cleaning functionality through the private method
-            // Since cleanBazelCommand is private, we test it indirectly through the parser
-            // The command should be cleaned when processed
-            
-            // Verify the original command contains the flags to be removed
-            XCTAssertTrue(originalCommand.contains("-Xfrontend"))
-            XCTAssertTrue(originalCommand.contains("-const-gather-protocols-file"))
-            XCTAssertTrue(originalCommand.contains("-emit-const-values-path"))
-            XCTAssertTrue(originalCommand.contains("-emit-module-path"))
-            XCTAssertTrue(originalCommand.contains("-index-store-path"))
-            XCTAssertTrue(originalCommand.contains("-index-ignore-system-modules"))
-            XCTAssertTrue(originalCommand.contains("-Xwrapped-swift"))
-            XCTAssertTrue(originalCommand.contains("-isysroot"))
-        }())
-    }
-    
-    // MARK: - Cache Tests
-    
-    func testBazelInterfaceCacheClearing() {
-        // Create valid workspace
-        let moduleFile = tempWorkspaceURL.appendingPathComponent("MODULE.bazel")
-        try! "module(name = \"test\")".write(to: moduleFile, atomically: true, encoding: .utf8)
-        
-        XCTAssertNoThrow(try {
-            let interface = try BazelInterface(workspaceRoot: workspacePath, bazelExecutable: "echo")
-            
-            // Cache clearing should not throw
-            interface.clearCache()
-        }())
-    }
-    
-}
